[BULLY] Fix race condition issue with ReShade

+++ src/hook.cpp
@@ -363,7 +363,7 @@ HRESULT CALLBACK Hook::hkGetDeviceState(IDirectInputDevice8* pThis, DWORD cbData
 			// Block camera rotation
 			mouseState->lX = 0;
 			mouseState->lY = 0;
-            io.MouseWheel = mouseState->lZ / WHEEL_DELTA;
+            io.MouseWheel = mouseState->lZ / static_cast<float>(WHEEL_DELTA);
 			mouseState->lZ = 0;
 
 			static int mouseCount = -1;
